+++ src/snowflake/cli/plugins/snowpark/package/commands.py
@@ -28,7 +28,7 @@ log = logging.getLogger(__name__)
 
 install_option = typer.Option(
     False,
-    "--install-from-pip",
+    "--pypi-download",
     help="Installs packages that are not available on the Snowflake anaconda channel.",
 )
 
@@ -45,17 +45,18 @@ deprecated_install_option = typer.Option(
 def package_lookup(
     name: str = typer.Argument(..., help="Name of the package."),
     install_packages: bool = install_option,
-    deprecated_install_packages :bool = deprecated_install_option,
-    allow_native_libraries: PypiOption = PackageNativeLibrariesOption,
+    deprecated_install_option: bool = deprecated_install_option,
+allow_native_libraries: PypiOption = PackageNativeLibrariesOption,
     **options,
 ) -> CommandResult:
     """
     Checks if a package is available on the Snowflake anaconda channel.
-    If the `--yes` flag is provided, this command checks all dependencies of the packages
+    If the `--pypi-download` flag is provided, this command checks all dependencies of the packages
     outside Snowflake channel.
     """
-    if deprecated_install_packages:
-        install_packages = deprecated_install_packages
+    if deprecated_install_option:
+        install_packages = deprecated_install_option
+
 
     lookup_result = lookup(
         name=name,
@@ -102,15 +103,15 @@ def package_create(
         help="Name of the package to create.",
     ),
     install_packages: bool = install_option,
-    deprecated_install_packages :bool = deprecated_install_option,
-    allow_native_libraries: PypiOption = PackageNativeLibrariesOption,
+    deprecated_install_option: bool = deprecated_install_option,
+allow_native_libraries: PypiOption = PackageNativeLibrariesOption,
     **options,
 ) -> CommandResult:
     """
     Creates a python package as a zip file that can be uploaded to a stage and imported for a Snowpark python app.
     """
-    if deprecated_install_packages:
-        install_packages = deprecated_install_packages
+    if deprecated_install_option:
+        install_packages = deprecated_install_option
 
     if (
         type(